Put SignalInfo, EncodingInfo in common module
SignalInfo, EncodingInfo are used across backends but were defined in soundfile backend and
imported from there in sox backend, which gives an impression that sox backend depends on sounds file backend.

This PR fixes this by moving these to a common module.
